% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGAS_preprocessing.R
\name{umap_coordinate}
\alias{umap_coordinate}
\title{scRNA-seq UMAP calculation}
\usage{
umap_coordinate(
  count,
  metadata = NULL,
  min.cells = 3,
  min.features = 200,
  nfeatures = 2000,
  dims = 1:10,
  resolution = 0.5
)
}
\arguments{
\item{count}{A count matrix or data frame with scRNA-seq data (rows = genes, columns = cells).}

\item{metadata}{metadata for scRNA-seq, default = NULL}

\item{min.cells}{filtering cells, default = 3}

\item{min.features}{filtering genes, default = 200.}

\item{nfeatures}{The number of high variable genes, default = 2000}

\item{dims}{Dimensions of reduction to use as input}

\item{resolution}{Value of the resolution parameter, use a value above (below) 1.0 if you want to obtain a larger (smaller) number of communities.}
}
\value{
A UMAP coordinate data.frame
}
\description{
Using Seurat to calculate UMAP corrdinate for scRNA-seq data for visualization
}
