% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGAS_preprocessing.R
\name{DEGAS_preprocessing}
\alias{DEGAS_preprocessing}
\title{DEGAS preprocessing}
\usage{
DEGAS_preprocessing(
  scst_list,
  patdata,
  phenotype,
  sclab = NULL,
  bulk_hvg = TRUE,
  bulk_de = TRUE,
  sc_de = TRUE,
  add_genes = NULL,
  n_hvg = 250,
  n_bulk_de = 250,
  n_sc_de = 200,
  padj.thresh = 0.05,
  model_type = "category"
)
}
\arguments{
\item{scst_list}{Count matrix or data frame with scRNA-seq/spatial transcriptomics data (rows = genes, columns = cells).}

\item{patdata}{Count Bulk RNA-seq data (genes × samples).}

\item{phenotype}{A phenotype vector or data frame describing sample labels as disease status.
If \code{model_type = "category"}, this should be a categorical variable.
If \code{model_type = "survival"}, it must include \code{time} and \code{status} columns.}

\item{sclab}{Metadata for scRNA-seq e.g. cell type, clusters (required), default = NULL}

\item{bulk_hvg}{Whether user wants to use sample (bulk) level high variable genes as input in DEGAS. Default is \code{TRUE}.}

\item{bulk_de}{Whether user wants to use sample (bulk) differential expression genes for phenotype using DEseq2, commonly set Normal or Control as reference. Default is \code{TRUE}.}

\item{sc_de}{Whether user wants to use single cell differential expression genes for each cluster. Default is \code{TRUE}.}

\item{add_genes}{The user decides whether to add their own gene list by preference. If the user set bulk_hvg, bulk_de and sc_de as FALSE, this gene list is necessary. Default is \code{NULL}.}

\item{n_hvg}{Number of bulk high variable genes, default = 250.}

\item{n_bulk_de}{Number of bulk differential expression genes for phenotype, default = 250.}

\item{n_sc_de}{Number of genes for each single cell clusters, default = 200.}

\item{padj.thresh}{Threshold for differential expression analysis, default = 0.05}

\item{model_type}{Choose model type (category or survival), default = \code{category}.}
}
\value{
A list ready for DEGAS analysis, includes SC/ST RNA-seq data, SC metadata, bulk RNA-seq data, bulk sample phenotype, SC/ST datalist name if multiple datasets.
}
\description{
Preprocess 4 necessary input data, scData, sclab, Patdata and phenotype with gene selection and normalization.
}
